Move sanity client to data util directory
